Add mousedown/up tracking to allow for change in behaviour in Chrome 73
Chrome 73 changes the behaviour of click events when the mousedown is on one element
and the mouseup is on another. The previous behaviour was to not emit any click event
at all. The new behaviour is to emit a click event with the target being the
common ancestor of the elements involved.

In the case of pickadate the calendar is popped up over the input element
meaning that the mousedown event targest the input but the mouseup targest
the calendar itself. This results in the click event targetting the common
ancestor which is not particularly useful.

The workaround here is to track the mousedown and mouseup events during
calendar opening and ignore any click events generated during this.

See https://bugs.chromium.org/p/chromium/issues/detail?id=946459&can=2&q=click
for additional details.